requirements.txt
# Core Framework
flask>=2.3.3

# Document Processing
PyPDF2==3.0.1
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=10.0.0

# AI and Language Processing
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=1.3.0
nltk>=3.8.1
rouge-score>=0.1.2

# Vector Database
chromadb>=0.4.22

# Data Processing and Visualization
numpy>=1.24.3
plotly>=5.16.1

# PII Detection and Anonymization
presidio-analyzer>=2.2.33
presidio-anonymizer>=2.2.33
spacy>=3.7.2

# Embeddings and ML
transformers>=4.31.0
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
tenacity>=8.2.2

backoff
sentencepiece

duckdb>=0.7.0